requirements.txt
paddlepaddle-gpu==2.2.2.post101
dapper==1.3.0
